refactor(notify): MCP通知を独立サイドカー oretachi-notify に分離
フックから GUI 本体(oretachi.exe)を --notify/--set-description で再起動して
いた CLI モードを撤去し、serde_json のみ依存の極小サイドダカー
oretachi-notify (src-tauri/notify) に分離。Tauri externalBin として同梱し、
hooks.json はサイドカーを直接呼ぶ。

- notify/: /notify と /set-description の両方を担うサイドカークレートを新設
  (HTTP送信を post_json に共通化、引数パース/stdin読取/server-info読取を移設)
- main.rs: CLI モードを撤去し run() のみに
- mcp_server.rs: standalone クライアント送信関数を削除(サーバ側ハンドラは保持)
- claude_plugin.rs: sidecar_path() を追加しフック command をサイドカーへ向ける
- tauri.conf.json: bundle.externalBin 追加・cli プラグイン削除、
  beforeBuildCommand で build:sidecar を実行
- scripts/build-sidecar.mjs: triple 検出→cargo build→binaries/ へステージング
- Cargo.toml: workspace 化(members=[notify])・tauri-plugin-cli 削除
- task_executor.rs: テストの WorktreeEntry に description フィールド補完
