+++ src-tauri/.gitignore
@@ -5,3 +5,6 @@
 # Generated by Tauri
 # will have schema files for capabilities auto-completion
 /gen/schemas
+
+# Staged sidecar binaries (generated by scripts/build-sidecar.mjs)
+/binaries/
