+++ src-tauri/Cargo.toml
@@ -7,6 +7,12 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+# 通知サイドカー (oretachi-notify) を同一ワークスペースのメンバーとして含める。
+# externalBin 用バイナリは scripts/build-sidecar.mjs が target/release から binaries/ へステージングする。
+[workspace]
+members = ["notify"]
+resolver = "2"
+
 [lib]
 # The `_lib` suffix may seem redundant but it is necessary
 # to make the lib name unique and wouldn't conflict with the bin name.
@@ -25,7 +31,6 @@ tauri-plugin-opener = "2"
 tauri-plugin-clipboard-manager = "2"
 tauri-plugin-dialog = "2"
 tauri-plugin-log = "2"
-tauri-plugin-cli = "2"
 tauri-plugin-global-shortcut = "2"
 tauri-plugin-os = "2"
 tauri-plugin-notification = "2"
