+++ src-tauri/capabilities/default.json
@@ -20,7 +20,6 @@
     "clipboard-manager:allow-read-text",
     "dialog:default",
     "log:default",
-    "cli:default",
     "global-shortcut:allow-register",
     "global-shortcut:allow-unregister",
     "global-shortcut:allow-is-registered",
