+++ src-tauri/notify/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "oretachi-notify"
+version = "0.20.3"
+description = "oretachi worktree notification sidecar"
+edition = "2021"
+
+[[bin]]
+name = "oretachi-notify"
+path = "src/main.rs"
+
+[dependencies]
+serde_json = "1"
