+++ src-tauri/src/claude_plugin.rs
@@ -66,6 +66,7 @@ pub fn generate_plugin_files(app_handle: &AppHandle) -> Result<(), String> {
     let exe_path = std::env::current_exe()
         .map(|p| p.to_string_lossy().replace('\\', "/"))
         .unwrap_or_else(|_| "oretachi".to_string());
+    let sidecar_path = sidecar_path();
 
     // plugin.json
     let plugin_json = build_plugin_json(&exe_path);
@@ -77,8 +78,8 @@ pub fn generate_plugin_files(app_handle: &AppHandle) -> Result<(), String> {
     )
     .map_err(|e| format!("Failed to write plugin.json: {}", e))?;
 
-    // hooks/hooks.json
-    let hooks_json = build_hooks_json(&exe_path);
+    // hooks/hooks.json （通知は GUI 本体ではなく oretachi-notify サイドカーが受ける）
+    let hooks_json = build_hooks_json(&sidecar_path);
     let hooks_json_path = hooks_dir.join("hooks.json");
     std::fs::write(
         &hooks_json_path,
@@ -162,15 +163,31 @@ fn build_plugin_json(exe_path: &str) -> serde_json::Value {
     })
 }
 
-fn build_hooks_json(exe_path: &str) -> serde_json::Value {
-    // exe_path はビルド時に確定した実行ファイルの絶対パスを直接ハードコードする。
+/// hooks.json の command が参照する通知サイドカー (oretachi-notify) の絶対パスを返す。
+/// externalBin としてバンドルされたサイドカーは GUI 本体と同じディレクトリに配置される
+/// （dev 時も target/debug 隣に複製される）ため、current_exe と同階層を指す。
+fn sidecar_path() -> String {
+    const SIDECAR_BIN: &str = if cfg!(target_os = "windows") {
+        "oretachi-notify.exe"
+    } else {
+        "oretachi-notify"
+    };
+    std::env::current_exe()
+        .ok()
+        .and_then(|p| p.parent().map(|d| d.join(SIDECAR_BIN)))
+        .map(|p| p.to_string_lossy().replace('\\', "/"))
+        .unwrap_or_else(|| "oretachi-notify".to_string())
+}
+
+fn build_hooks_json(notifier_path: &str) -> serde_json::Value {
+    // notifier_path は通知サイドカー (oretachi-notify) の絶対パスを直接ハードコードする。
     // ${user_config.XXX} は Claude Code プラグイン仕様のユーザー設定値展開構文。
     // 各ワークツリーの pluginConfigs.oretachi@oretachi.options から値が注入される。
     let mut hooks = serde_json::Map::new();
     for (event, key) in EVENT_CONFIG_KEYS {
         let command = format!(
             "\"{}\" --notify \"${{user_config.worktree_name}}\" --kind ${{user_config.{}}} --agent cc",
-            exe_path, key
+            notifier_path, key
         );
         hooks.insert(
             event.to_string(),
@@ -181,15 +198,15 @@ fn build_hooks_json(exe_path: &str) -> serde_json::Value {
         );
     }
 
-    // ExitPlanMode フック: プラン確定時に oretachi exe を起動し、プランを AI 要約して
-    // ワークツリーの description にセットする。
+    // ExitPlanMode フック: プラン確定時に通知サイドカー (oretachi-notify) を起動し、
+    // プランを AI 要約してワークツリーの description にセットさせる。
     // ExitPlanMode はユーザー操作を伴うツールのため PreToolUse/PostToolUse は発火せず、
     // PermissionRequest イベントで発火する（matcher "ExitPlanMode"、プラン本文は tool_input.plan）。
     // 既存の通知用 PermissionRequest グループ(matcher "")と共存させるため配列に追記する。
     // decision は返さず観測のみ（exit 0）なので通常の承認フローはブロックしない。
     let set_desc_command = format!(
         "\"{}\" --set-description \"${{user_config.worktree_name}}\"",
-        exe_path
+        notifier_path
     );
     let exit_plan_group = serde_json::json!({
         "matcher": "ExitPlanMode",
