+++ src-tauri/src/mcp_server.rs
@@ -1994,139 +1994,3 @@ pub fn start_mcp_server(app_handle: AppHandle, port: u16, remote_access: bool) {
         let _ = complete_tx.send(());
     });
 }
-
-// ─── CLI notification sender (standalone, no AppHandle needed) ───────────────
-
-pub fn send_notification_standalone(worktree_name: &str, kind: Option<&str>, body: Option<&str>, agent: Option<&str>) -> Result<(), String> {
-    let (port, api_key) = read_server_info_standalone()?;
-    let mut payload = serde_json::json!({
-        "worktree": worktree_name,
-        "kind": kind.unwrap_or("general"),
-    });
-    if let Some(b) = body {
-        payload["body"] = serde_json::Value::String(b.to_string());
-    }
-    if let Some(a) = agent {
-        payload["agent"] = serde_json::Value::String(a.to_string());
-    }
-    let payload_str = payload.to_string();
-    let request = format!(
-        "POST /notify HTTP/1.1\r\nHost: 127.0.0.1:{port}\r\nContent-Type: application/json\r\nAuthorization: Bearer {api_key}\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{payload_str}",
-        payload_str.len()
-    );
-
-    use std::io::Write;
-    use std::time::Duration;
-
-    let addr: std::net::SocketAddr = format!("127.0.0.1:{}", port)
-        .parse()
-        .map_err(|e| format!("Invalid address: {}", e))?;
-    let mut stream = std::net::TcpStream::connect_timeout(&addr, Duration::from_secs(3))
-        .map_err(|e| format!("Cannot connect to oretachi MCP server: {}", e))?;
-    stream
-        .set_write_timeout(Some(Duration::from_secs(5)))
-        .map_err(|e| format!("Failed to set write timeout: {}", e))?;
-    // 短い読み取りタイムアウトで応答をチェック（非ブロッキング性を維持しつつ確実な配信失敗を検出）
-    stream
-        .set_read_timeout(Some(Duration::from_millis(500)))
-        .map_err(|e| format!("Failed to set read timeout: {}", e))?;
-    stream
-        .write_all(request.as_bytes())
-        .map_err(|e| format!("Failed to send notification: {}", e))?;
-    stream
-        .flush()
-        .map_err(|e| format!("Failed to flush: {}", e))?;
-
-    // 応答の最初の行だけ読んでステータスを確認する（タイムアウトは無視してベストエフォートとする）
-    use std::io::{BufRead, BufReader};
-    let reader = BufReader::new(&stream);
-    if let Some(Ok(first_line)) = reader.lines().next() {
-        if first_line.starts_with("HTTP/") && !first_line.contains(" 200 ") {
-            return Err(format!("Server returned unexpected response: {}", first_line));
-        }
-    }
-    // タイムアウトや読み取りエラーはベストエフォート成功扱い（フックのブロック防止）
-    Ok(())
-}
-
-pub fn send_set_description_standalone(worktree_name: &str, hook_json: Option<&str>) -> Result<(), String> {
-    let (port, api_key) = read_server_info_standalone()?;
-    let mut payload = serde_json::json!({
-        "worktree": worktree_name,
-    });
-    if let Some(j) = hook_json {
-        payload["hookJson"] = serde_json::Value::String(j.to_string());
-    }
-    let payload_str = payload.to_string();
-    let request = format!(
-        "POST /set-description HTTP/1.1\r\nHost: 127.0.0.1:{port}\r\nContent-Type: application/json\r\nAuthorization: Bearer {api_key}\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{payload_str}",
-        payload_str.len()
-    );
-
-    use std::io::Write;
-    use std::time::Duration;
-
-    let addr: std::net::SocketAddr = format!("127.0.0.1:{}", port)
-        .parse()
-        .map_err(|e| format!("Invalid address: {}", e))?;
-    let mut stream = std::net::TcpStream::connect_timeout(&addr, Duration::from_secs(3))
-        .map_err(|e| format!("Cannot connect to oretachi MCP server: {}", e))?;
-    stream
-        .set_write_timeout(Some(Duration::from_secs(5)))
-        .map_err(|e| format!("Failed to set write timeout: {}", e))?;
-    stream
-        .set_read_timeout(Some(Duration::from_millis(500)))
-        .map_err(|e| format!("Failed to set read timeout: {}", e))?;
-    stream
-        .write_all(request.as_bytes())
-        .map_err(|e| format!("Failed to send set-description: {}", e))?;
-    stream
-        .flush()
-        .map_err(|e| format!("Failed to flush: {}", e))?;
-
-    use std::io::{BufRead, BufReader};
-    let reader = BufReader::new(&stream);
-    if let Some(Ok(first_line)) = reader.lines().next() {
-        if first_line.starts_with("HTTP/") && !first_line.contains(" 200 ") {
-            return Err(format!("Server returned unexpected response: {}", first_line));
-        }
-    }
-    Ok(())
-}
-
-fn read_server_info_standalone() -> Result<(u16, String), String> {
-    #[cfg(target_os = "windows")]
-    let base = {
-        let appdata = std::env::var("APPDATA")
-            .map_err(|_| "APPDATA environment variable not set".to_string())?;
-        PathBuf::from(appdata).join("com.ia.oretachi")
-    };
-
-    #[cfg(not(target_os = "windows"))]
-    let base = {
-        let home = std::env::var("HOME")
-            .map_err(|_| "HOME environment variable not set".to_string())?;
-        PathBuf::from(home)
-            .join("Library")
-            .join("Application Support")
-            .join("com.ia.oretachi")
-    };
-
-    // mcp-server.json を優先して読む
-    let json_path = base.join(SERVER_INFO_FILE);
-    if json_path.exists() {
-        let content = fs::read_to_string(&json_path)
-            .map_err(|e| format!("Cannot read server info file: {}", e))?;
-        let info: serde_json::Value = serde_json::from_str(&content)
-            .map_err(|e| format!("Invalid server info JSON: {}", e))?;
-        let port = info["port"].as_u64()
-            .ok_or_else(|| "Missing port in server info".to_string())? as u16;
-        let api_key = info["apiKey"].as_str()
-            .ok_or_else(|| "Missing apiKey in server info".to_string())?
-            .to_string();
-        return Ok((port, api_key));
-    }
-
-    // mcp-server.json が存在しない場合、APIキーが取得できないためエラーとする
-    Err("Cannot read API key: mcp-server.json not found. Please restart oretachi to regenerate the server info file.".to_string())
-}
