+++ src-tauri/tauri.conf.json
@@ -6,7 +6,7 @@
   "build": {
     "beforeDevCommand": "pnpm dev",
     "devUrl": "http://localhost:1420",
-    "beforeBuildCommand": "pnpm build",
+    "beforeBuildCommand": "pnpm build && pnpm build:sidecar",
     "frontendDist": "../dist"
   },
   "plugins": {
@@ -15,16 +15,6 @@
         "https://github.com/ishida-supsys/oretachi/releases/latest/download/latest.json"
       ],
       "pubkey": "dW50cnVzdGVkIGNvbW1lbnQ6IG1pbmlzaWduIHB1YmxpYyBrZXk6IDMxQjBDQTZBOEUyNjBDRDEKUldUUkRDYU9hc3F3TVhwbFhFTlVEb0lJcVJENjlZMENuRVlRSjkvalIrakZLbW9qT1VpOGZ5cGoK"
-    },
-    "cli": {
-      "args": [
-        {
-          "name": "notify",
-          "short": "n",
-          "description": "通知するワークツリー名",
-          "takesValue": true
-        }
-      ]
     }
   },
   "app": {
@@ -52,6 +42,9 @@
     "publisher": "ishida-supsys",
     "shortDescription": "Multi-session terminal manager for developers",
     "longDescription": "oretachi is a desktop application for managing multiple terminal sessions with a tabbed interface, split-pane layout, worktree support, and tray notifications to keep your development workflow organized.",
+    "externalBin": [
+      "binaries/oretachi-notify"
+    ],
     "icon": [
       "icons/32x32.png",
       "icons/128x128.png",
